// src/web/src/api/generated/models/WeeklyEvaluationSummaryResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WeeklyEvaluationSummaryResponse = {
    id: string;
    weekStart: string;
    weekEnd: string;
    formatVersion: string;
    planVersionId: string | null;
    cutoffAt: string;
    status: string;
    trafficLight: string;
    rationale: string;
    createdAt: string;
    hasDecision: boolean;
};